Separate Windows partition growth from device I/O
Move partition discovery, validation, growth, and verification into a
dedicated Windows source file. Keep device access, locking, and block
I/O in the device backend, with private helpers for shared diagnostics.
