+++ CMakeLists.txt
@@ -136,6 +136,7 @@ if(EXFAT_RESIZE_BUILD_CLI)
             src/windows/device.c
             src/windows/device_path.c
             src/windows/entry.c
+            src/windows/partition.c
         )
     else()
         set(
